Adds main function
this commit adds the main function, which takes in many parameters and
creates the corresponding stimuli, skipping duplicates and empty
stimuli. it takes the largest mask among all created stimuli and applies
that to each image (so that each has the same mask and none show any
aliasing), returning both the masked and unmasked stimuli. some
parameters (size, origin, and number_of_fade_pixels) are required to be
the same across all images, but alpha, the frequencies, phase, and
amplitude are allowed to vary.