adds check_ and plot_stim_properties
these two functions allow the user to visualize some properties of the
stimuli that they're probably interested in: the mask radius and the
min/max of spatial frequency (in both cycles per pixel and cycles per
degree, with and without mask). these values are calculated and then
stored in a dataframe and are visualized using heatmaps on a
facetgrid. it can take a while to fully explore the space, but this is a
useful demonstration