main uses output dir, adds command-line function
with this, stimuli now uses output_dir instead of filename (and so
always saves out as run__.npy in that direcotry, as well as the
unshuffled one). it also (should) correctly use the blank trials now,
though this needs more testing. it also can be run from the command
line, though the functionality is very basic right now