adds bytescaling
the reason the stimuli presentation was so slow and the saved npy files
were so large is that the arrays were doubles (float64, I think), which
meant they were huge. However, for presentation we want them to be 0 to
255 values, uint8 (that is, one byte). so we make use of the scipy
function bytescale, which handles this for us. it only gets applied in
gen_stim_set, after the application of the mask, and is the defualt for
main