changes how timing is handled
previously, we were looping and calling .draw and .flip once per
frame. Jon and Noah think this is not the right way to do this, so now
we loop through the stimuli, calling .flip only when we need to put a
new stimuli on or off. this cleans things up considerably. things work
as written, but considerable cleanup is needed