Changes how stimuli randomization handled
This commit changes how we do stimuli randomization. previously, we
would save a different .npy file for each run, each one randomized
differently (along with an unshuffled version). now, we save one
unshuffled version and a bunch of indices (1d arrays saved as .npy
files) for the number of runs we want. their path also depends on the
subject name. additionally, we ensure that there are never two blanks in
a row and we save a pandas dataframe (as a csv) that contains all the
frequency / phase / alpha, etc information of the unshuffled data. since
we'll want to create indices more often than we create stimuli, adds a
create_stim flag, which, when false, only creates the stimuli and
dataframe.

this also updates the experiment script then to reflect this. run now
takes both a stimuli and index path arg and will return the
indices. expt loads in the idx_paths itself, based on the subject name
and number of runs. both paths, plus the index, are saved in the
h5py. we also have a quick bugfix: if a variable is None, we wrap it in
str to make a dataset out of it. the command line parser is also updated
to reflect this, taking both number_of_runs and subj_name as
regular (non-flag) arguments, and input and output dir as flag /
optional (they have default values which should work).