change digit to every other stimuli
previously, we were presenting the digit during the stimuli ON block and
a blank / delay during the stimuli OFF block. this ended up being very
fast. now, we show a digit for a stimuli ON and OFF blocks, then a blank
for the next stimuli ON and OFF blocks. this makes the task much more
manageable.