stimuli.main now accepts create_idx kwarg
similar to create_stim, when this flag is true we save the permuted
indices, when it's false we skip that whole process.

the descriptive dataframe is now only created and saved when the stimuli
are.

we now also check for existence before saving, for the stimuli,
descriptive dataframe, and indices. if these exist, we throw an
exception and quit out.