adds outer mask to stimuli
with this commit, we update _fade_mask to be able to handle masks with
both inner and outer radii, add a create_outer_mask function, and have
gen_stim_set (and thus main) make use of this new function when creating
the super mask. create_mask has been renamed to create_antialiasing_mask
to better reflect what it does.