adds scatter_ci_dist
this function is meant to be used with the bootstrap data that
GLMdenoise returns and will accurately draw the CIs using
them. everything that gets handed to this (again, to be used with
sns.FacetGrid.map_dataframe) will be assumed to correspond to one point
and all the values will be assumed to be bootstraps of whatever you want
to plot. therefore, you simply plot the mean and take the 2.5th and
97.5th percentile from that data to use as the endpoints of the CI.