Adds hemisphere averaging to first level analysis
with this, adds an option which collapses ROIs across hemispheres. this
does nothing if eccen_bin is False (that is, if we don't bin
eccentricity into 1 degree chunks; since in that case we don't know what
parts of the left and right hemispheres correspond to each other), but
otherwise this averages together corresponding ROIs, since we often
don't want to consider them separately. I leave the option to keep them
separate since they often *do* differ from each other, so we may want to
look into that in the future.

also makes sure that this works both with eccen_bin set to True and
False, which I hadn't tested in a while.