Adds local spatial frequency calculation
Small cluster of related changes here, all based around adding the local
spatial frequency calculation into first_level_analysis.py. With this
new function, we calculate the local spatial frequency in the stimuli as
part of the creation of the results dataframe. However, currently this
is incomplete: it only works for circular stimuli; need to add bit to
handle this for other stimuli later.

 - design_df now includes res and alpha, since these contain necessary
   information for computing the local spatial frequency info

 - stim_rad_deg, the radius of the stimulus in degrees of visual angle,
   is now a necessary input to the create_GLM_result_df function

 - corrects the default value for diameter of the whole screen from 28
   to 24

 - renames max_degree_rad in some utils functions to stim_rad_deg, since
   it's the same info and I feel like stim_rad_deg is a little more
   intuitive of a name