Adds find_ecc_range functions
This commit adds two functions to first_level_analysis, both of which
return information about the extent of the screen that the stimulus is
presented on. The first function does this in pixels, the second does it
in degrees. Both use ppt.mkR to find the minimum / maximum distance from
the origin where a given stimulus does not have value 127 (so this
assumes that is the "null value", not 0), and the degrees one just
converts this from pixels to visual angle using the radius of the
stimulus in degrees