Splits third notebook in half
This commit splits the first level analysis notebook into two parts: one
about the creation of the design matrix and one about analyzing the
results of the GLM analysis, which should be run between the two. This
follows from splitting up the first_level_analysis script in two.

This commit also updates the text in the README and Design-matrices
notebook to reflect the creation of the design_matrices.py script.

The 04-MRI-first-level notebook currently contains a bunch of extra
stuff about local spatial frequency that will be moved out