Corrects local spatial frequency calculation
with this commit, we correct the local spatial frequency calculation and
make it work for all types of stimuli. The values used are now derived
analytically and so are correct and work for stimuli types. While this
changes the functions in stimuli.py (which are used to determine where
aliasing occurs), it doesn't change the masks in any noticeable
way (though they may be off by a pixel or so, it's hard to tell)