Adds plotting of linear approximations
With this commit, the functions that plot the linear approximations of
the gratings are moved into the scripts and out of the notebooks. They
are local_grad_sin (to create the local sin grating based on the
gradients in cycles / pixel) and plot_grating_approximation (which
creates the windowed version of the grating and the linear
approximation). Additionally:

 - mask_array_like_grating has been added. this function takes in two
   arrays and masks the second like the first. it is now being used in a
   couple of functions

 - create_sin_cpd functions as before, but now uses create_sin_cpp,
   which creates a full 2d sine wave, with frequency specified in cycles
   / pixel. this is because we want the frequencies in cpp for the
   linear approximation plots

 - create_sf_maps_cpp now makes sure that size is an int (because it
   fails if it's a float)