bugfix: adds catch
this adds a little bit to catch a weird bug I was having. if the
unshuffled stimuli description csv is smaller than the behavioral
results shuffled indices (which happened when shifting to the new type
of stimuli: the new version has fewer stimuli than the old version),
then the code I have will be unable to determine the size of each
stimulus class but will also not throw an exception of any kind. this is
bad, so this custom exception avoids that