gets first_level_analysis rule working
the issue I was having was due to my change in how nibabel reports the
shape of mgz files. My "3d" mgz files are actually 4d ones, just with
a final singleton dimension that nibabel was ignoring. so in my
first_level_analysis code, I wasn't properly squeezing the
newly-4d-but-actually-only-1d-of-data files. the little fix handles
that.

other changes:

 - the command-line class_nums argument is now a single integer and
   runs classes in range(class_nums) instead of a list. the main
   class_nums argument is still a list.

 - adds stim_type command line argument.

 - df.varea gets cast to an int

 - adds stim_type param to the first_level_analysis Snakefile rule.

 - figured out how to get the dynamic input working. it's kind of
   hacky, but now I require all models_class*mgz files for both
   df_mode summary and full, even though they're actually only needed
   for full. I couldn't get dynamic working within a function and
   using expand with 48 / 52 would lag out.

 - adds resources to rule

 - adds warning if we're not binning anything

 - adds first_level_all rule