updates how save_results_niftis works
with this commit, save_results_niftis no longer has dynamic outputs,
but is called once per model class. dynamic is weird and made things
take a really long time when constructing the DAG. the output is also
not *really* dynamic: we know exactly how many there will be, it just
depends on the session. so we handle this similarly to how we handled
preprocessing, calling the rule once per model class / run and having
an input function on the next rule that determines how many there
should be