Updates Snakefile for new preprocessing
I just changed how preprocessing handles the output directory when the
bids structure is BIDS, so this changes things to work with that. the
big difference is `-outdir` is no longer an argument: we want to place
this in a reasonable directory, so the user can't completely specify
the output anymore. we can, however, specify the directory under
preprocessed and the suffix before .nii.gz, and we make use of the
first of those here.