adds plot_data to plotting
with this commit, adds new function, plot_data to plotting.py. this
function plots the amplitude estimates (with 68% confidence interval, by
default) of the data as a function of the freq_space_distance (by
default, can attempt to do it with anything). as part of this, adds
ability to call this function from the command line and also moves a
whole bunch of plotting helper functions from utils.py to plotting.py