Adds ability to model blanks in design matrix
with this commit, user can now model blanks in the stim_class design
matrix. user must specify how many blanks (1 through 10, inclusive) to
model and then these will be added as another trial type / models
class (with the max number). in the BIDS events tsv, we will also add a
note (in the new notes column) which explains that these are "blank
trials".

 - also adds full_TRs argument to create_BIDS_tsv.main, allowing us to
   correctly check the length of the runs for both pilot and non-pilot
   runs