adds summary_plots
this commit adds the summary_plots script and associated rules, which
summarize the tuning curve parameters. note that the defaults probably
won't be great if there's an outlier so not sure what to do about that.

this also removes the confusingly named "tuning_curve_peak_octaves". if
you want to get at this, use a semilogy or loglog plot