increases blank times
we're having trouble estimating the baseline, so we add 16 seconds of
blank at the beginning of each run. we also add 8 more seconds at the
end of each run.

also, updates the neutral color to 128, as our new stimuli use that
value. this didn't seem to be noticeable for the previous ses-01, ses-02
scans, but who knows.