adds benson_template_names argument to first level
because we want to include the "sigma" benson retinotopy result, but
it's not there for every subject's priors. this way, we allow priors and
posteriors to have different bunches of benson_template_names and this
should make it easier to add any new ones, if they get added

also, this makes the results_template_path and benson_template_path
flagged args (though still required), which makes it easier to tell
where benson_template_names ends and those begin. I guess I could have
just reordered things, but this seems easier.