updates period_summary_plot
period_summary_plot now works differently. instead of only working on
circular and radial, it now works on all stimulus superclasses. and
instead of assuming all are fit with lines through the origin, all are
allowed to have non-zero intercepts (which means that we use a slightly
different log-polar stimulus for each eccentricity)