adds single-point spatial frequency funcs
previously, the only way to get the spatial frequency was to create a
full image, which meant you needed to grab that and then find the pixel
closest to the pRF center in order to get the local spatial
frequency. with this commit, adds the ability to calculate the spatial
frequency for a single location (given an eccentricity and angle), which
is *much* faster.

with all this, need to do a big update to first_level_analysis to
overhaul how spatial frequency is gotten into the dataframe and to
change binning

other changes:

 - create_sf_maps_cpd now also returns dx, dy, and direction (not just
   the magnitude)

 - updates function calls to create_sf_maps_cpd to reflect this change

 - adds the ability (for both maps and single-values) to return the
   spatial frequency in "origin-referenced polar spatial frequency";
   that is, in the radial and angular directions.

 - updates the model notebook:

   - adds log-Gaussian donut model (in addition to the regular 2d
     gaussian)

   - looks at both x-y and w_r-w_a space

   - adds transformation of the angle from the values from Noah's
     retinotopy to the values needed for the origin-referenced polar
     calculation