renames stimulus superclasses
to get them correct and in-line with how I presented them in VSS. What I
was referring to as "circular" is now "radial", "radial" is now
"angular", "horizontal" is now "vertical", and "vertical" is now
"horizontal". the label now describes which frequency is non-zero,
bringing it in-line with the practice in systems neuroscience.

will still need to relabel this in the notebooks, but will need to
regenerate the csvs first