Adds ability to not load in results_df
in analyze_model.py, adds the ability to not load in results_df (only
load in loss and model parameters). depending on your number of voxels,
these results dataframes may be too large to comfortably load into
memory. this enables us to only load in the loss and model parameters,
which are generally what we're comparing anyway