bugfix in how Dataset gets voxels
when there's a df_filter so that the Dataset does not include all
voxels, iterating through the dataloader threw an exception as it tries
to find a voxel that has been dropped. this changes how __getitem__
works by adding a new column which just reindexes the remaining
columns. the functionality of getting a specific voxel still remains,
with the return of the function get_voxel