changes up how the model is structured
this commit has a bunch of changes to it. we end up overhauling a lot of
how the LogGaussianDonut is constructed, in order to allow and make
easier the following: the full model now has 11 parameters, since it can
train the amplitude and mode parameters for both relative and absolute
reference frames; can now decide whether to train the amplitude
parameters or not (separate from the mode parameters). in order to do
this, I removed all the train_* kwargs from the initialization call and
instead based that call around three model specification kwargs: how to
handle eccentricity (scaling, constant, or full), how to handle
orientation (absolute, relative, iso, or full), and whether to train the
amplitude parameters (True or False). this then determines how
everything else is set.

other changes:

 - LogGaussianDonut.__str__() now shows the different parameter names,
   not just their values. I'm not sure I like this, because it's busy,
   so I might change it

 - every LogGaussianDonut calculates the relative_sf_angle

 - removes the different wrapper classes as children of
   LogGaussianDonut; now they're handled by just setting those three
   model specification kwargs

 - change how called from command line from model_type to the three
   model spec kwargs

 - changes Snakefile rules to acount for that