fixes bytescale issues
bytescale only shows up in sklearn after version 0.19.1, but that's
the version that installing pytorch via conda downgrades you to. so I
copied over the function into utils myself. however, this doesn't fix
the issue with experiment.py, since that runs from a separate
environment and so can't import from any of the other scripts. to do
that, I just hard coded it, since it's pretty straightforward