bugfix
for some reason, the reshape I was doing on the data stored in the
GLMdenoise results mat was rearranging the values when the array was 4
or 5d (i.e., models, modelmd, and modelse), making it impossible to
line up those values with those of the 1d ones (R2, varea, eccen,
etc). switching to using squeeze and, when necessary, transpose, seems
to fix the issue.