raise exception if nan prediction,removes bad test loss calculation
we were occasionally (during cross-validation) getting a nan during
prediction and kept going, which could result in nan parameters, which
would mess everything up. that doesn't happen everytime, so we raise
an exception and rely on snakemake to then rerun the job, and
hopefully get lucky the second time

we also remove the old way of doing the test loss calculation, which
simplifies the code a lot.