adds correct sigma to first level results
after talking with Noah, it turns out we don't want to use the sigma
mgz created by Bayesian retinotopy (which are just the slope from
Kay2013 times the eccentricity of the voxel), but instead the outputs
of the pRF fit

as part of this, I also went ahead and rearranged where those Bayesian
retinotopy outputs are stored. I had been putting them in the
freesurfer directory, but now have them in a separate derivative,
prf_solutions, which has separate sub-folders for data (the pRF fits),
atlas (the Benson14 atlas, which serves as a prior for Bayesian
retionotopy), and bayesian_posterior (the posterior of the Bayesian
retinotopy). this required a bit of restructuring of both Snakefile
and first_level_analysis.py

others changes:

 - the atlas_type snakemake wildcard is now the same as the sub-folder
   from derivatives/prf_solutions that we grab those values from. so
   we have atlas and bayesian_posterior instead of prior and posterior

 - first_level_analysis now expects benson_template_path to have three
   %s symbols in it

 - first_level_analysis accepts two new arguments,
   prf_data_names (which tells it which mgzs to grab from the data
   sub-folder) and benson_atlas_type (which tells it which
   prf_solutions sub-folder to look for benson_template_names in)

 - benson_template_names and prf_data_names can both have more than
   just the variable name (angle, eccen, etc) in them, e.g.,
   inferred_, benson14_, full-, or all00- can also be included (and
   probably needs to be, in order to use benson_template_path to work
   for both the benson_template_names and prf_data_names)

 - we also load in the prf variance explained, and rename the columns
   GLM_R2 and prf_vexpl to make it clear the difference there