adds many summary rules to Snakefile
in order to make my life easier, this creates rules meant to create
helpful intermediary / final outputs (not the figures but the step
right before that). there's a bit of messiness because we need a
version for the simulated and for the actual, and there might be some
unnecessary duplication between cross-validated and not.