adds figures for different parts of the visual field
in addition to model fit on all voxels in V1, also fit it to different
parts: left half, right half, upper half, etc. we can now make (some
of) the params and features plots for these restricted models. we
can't make those plots that require having access to all the
bootstraps, but we can for all the others

also replaced the feature's plot_kind possibilities from all and
subject to median and bootstraps, which is clearer