updates environment-psychopy.yml
old version of environment-psychopy.yml was not working, but this one
does (it was running into conflicts that conda couldn't resolve). I
will create a docker image based on this one to avoid that issue

also updates README and adds some docs to Snakefile to explain a bit
about display and generating stimuli.