adds rule to add visual relative legend to figure
right now, just works with the pref-period bootstraps-overall angles-avg
feature figure, but could be extended for more

we do this by saving a svg of the legend and then using svgutils.compose
to add it on top. it is possible to do something similar using
matplotlib, but it was too hard to get the spacing correct. this way we
have a version with the spacing we want and we just scale it

other small changes:

 - removes title and subject=='all' from this feature figure

 - replace 'dpc' with 'deg'