enlarge panel label, rename to relative amplitude
with this, we rename max amplitude (in feature df figure) to relative
amplitude, which will hopefully be a bit clearer. we also make the panel
labels larger, so they're not the same size as the other text. with
these changes, some minor adjustments to position are required in the 2d
summary figure