fix to get it working with sbatch
the path we were giving to sbatch for the location of its output and
error logs was the singularity-internal one, so it was throwing an
error. to fix that, bind DATA_DIR to the same location in the
container, so the paths are identical. do the same thing for
WORKING_DIR